+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+WIP version 2.0.0:
+------------------
+
+  - Add support for remote snapshot mode that will be available with powa 4
+  - Add pg_track_reboot_log function
+
 2018-07-15 version 1.0.1:
 -------------------------